Move platform specific logic for tty to different files
In order to support the windows platform, we will split the platform
specific logic into the tty directory.
