+++ src/tty/common.rs
@@ -1,0 +1,7 @@
+extern crate libc;
+
+/// Check to see if `fd` is a TTY
+pub fn is_a_tty(fd: libc::c_int) -> bool {
+    unsafe { libc::isatty(fd) != 0 }
+}
+
