+++ src/tty/windows.rs
@@ -1,0 +1,5 @@
+
+/// Checking for an unsupported TERM in windows is a no-op
+pub fn is_unsupported_term() -> bool {
+    false
+}
